Reverted back to original read_length
Read length should be good, and also adding test for image verification

+++ mio/devices/gs/config.py
@@ -37,9 +37,8 @@ class GSDevConfig(StreamDevConfig):
         i.e. the list will be the length of the number of buffers per frame,
         and each item in the list is the number of pixels in that buffer
         """
-
         total_pixels = self.frame_width_input * self.frame_height
-        buffer_npix = [self.max_pixels_per_buffer] * np.ceil(total_pixels / self.max_pixels_per_buffer)
+        buffer_npix = [self.max_pixels_per_buffer] * int(np.ceil(total_pixels / self.max_pixels_per_buffer))
         remainder = total_pixels % self.max_pixels_per_buffer
         if remainder != 0:
             buffer_npix[-1] = remainder
